ocr-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
paddleocr==2.9.1
paddlepaddle==2.6.2
pillow==11.0.0
python-multipart==0.0.17
numpy==1.26.4